refactor: cloud card unique url
