--- src/routes/card/+page.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-    import Display from './display.svelte';
-
-    export let data;
-</script>
-
-<Display userId={data.account.$id} variant="owner" />
