+++ src/routes/card/[uid]/+page.ts
@@ -1,25 +1,33 @@
 import { VARS } from '$lib/system.js';
 import { redirect } from '@sveltejs/kit';
-import { getCardImgUrls } from '../helpers.js';
 
-export const ssr = false;
+function getCardImgUrls(userId: string, endpoint: string) {
+    const resolved = endpoint;
 
-function urlContentToDataUri(url) {
-    return fetch(url)
-        .then((response) => response.blob())
-        .then(
-            (blob) =>
-                new Promise((callback) => {
-                    const reader = new FileReader();
-                    reader.onload = function () {
-                        callback(this.result);
-                    };
-                    reader.readAsDataURL(blob);
-                })
-        );
+    const frontImg = `${resolved}/cards/cloud?userId=${userId}`;
+    const backImg = `${resolved}/cards/cloud-back?userId=${userId}`;
+    const ogImg = `${resolved}/cards/cloud-og?userId=${userId}`;
+
+    return { frontImg, backImg, ogImg };
 }
 
-export async function load({ params, url }) {
+async function urlContentToDataUri(url: string): Promise<string> {
+    const res = await fetch(url);
+    const blob = await res.blob();
+
+    return await new Promise((callback) => {
+        const reader = new FileReader();
+        reader.onload = function () {
+            if (typeof this.result !== 'string') return '';
+            callback(this.result);
+        };
+        reader.readAsDataURL(blob);
+    });
+}
+
+export async function load({ params, url, parent }) {
+    const { account } = await parent();
+
     const userId = params.uid;
     const endpoint = VARS.APPWRITE_ENDPOINT ?? `${url.origin}/v1`;
     const { frontImg, backImg } = getCardImgUrls(userId, endpoint);
@@ -32,6 +40,14 @@ export async function load({ params, url }) {
 
     return {
         userId: params.uid,
-        frontBase64: await urlContentToDataUri(frontImg)
+        base64: {
+            front: await urlContentToDataUri(frontImg),
+            back: await urlContentToDataUri(backImg)
+        },
+        isOwner: account && account.$id === userId,
+        imgUrls: {
+            frontImg,
+            backImg
+        }
     };
 }
